Initialize filters with the first point, not (0,0)
The exp and predictive filters get a slight advantage since they are
initialized with the point (0,0), which is the same as the actual
first noise-free point. Instead we should initialize them with the
value of the first point, since that is the only sensible prediction
at that point in time.